third_party/WebKit/LayoutTests/fast/dom/CSSStyleDeclaration/css-style-declaration-named-getter.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body style="border: 1px; -webkit-box-sizing: border-box">
<script>

shouldBeUndefined("document.body.style['BorderTopWidth']");
shouldBeEqualToString("document.body.style['borderTopWidth']", "1px");
shouldBeUndefined("document.body.style['border-topWidth']");
shouldBeUndefined("document.body.style['border-TopWidth']");
shouldBeUndefined("document.body.style['borderTop-width']");
shouldBeUndefined("document.body.style['borderTop-Width']");
shouldBeEqualToString("document.body.style['border-top-width']", "1px");
shouldBeUndefined("document.body.style['CssBorderTopWidth']");
shouldBeUndefined("document.body.style['cssBorderTopWidth']");
shouldBeUndefined("document.body.style['cssborderTopWidth']");
shouldBeUndefined("document.body.style['cssBorder-top-width']");
shouldBeUndefined("document.body.style['css-border-top-width']");

shouldBeEqualToString("document.body.style['WebkitBoxSizing']", "border-box");
shouldBeEqualToString("document.body.style['webkitBoxSizing']", "border-box");
shouldBeUndefined("document.body.style['webkit-boxSizing']");
shouldBeUndefined("document.body.style['webkit-BoxSizing']");
shouldBeUndefined("document.body.style['webkitBox-sizing']");
shouldBeUndefined("document.body.style['webkitBox-Sizing']");
shouldBeEqualToString("document.body.style['-webkit-box-sizing']", "border-box");
shouldBeUndefined("document.body.style['cssWebkitBorderBoxSizing']");

</script>
</body>
</html>
