third_party/blink/web_tests/platform/generic/fast/dom/CSSStyleDeclaration/css-style-declaration-named-getter-expected.txt

PASS document.body.style['BorderTopWidth'] is undefined.
PASS document.body.style['borderTopWidth'] is "1px"
PASS document.body.style['border-topWidth'] is undefined.
PASS document.body.style['border-TopWidth'] is undefined.
PASS document.body.style['borderTop-width'] is undefined.
PASS document.body.style['borderTop-Width'] is undefined.
PASS document.body.style['border-top-width'] is "1px"
PASS document.body.style['CssBorderTopWidth'] is undefined.
PASS document.body.style['cssBorderTopWidth'] is undefined.
PASS document.body.style['cssborderTopWidth'] is undefined.
PASS document.body.style['cssBorder-top-width'] is undefined.
PASS document.body.style['css-border-top-width'] is undefined.
PASS document.body.style['WebkitBoxSizing'] is "border-box"
PASS document.body.style['webkitBoxSizing'] is "border-box"
PASS document.body.style['webkit-boxSizing'] is undefined.
PASS document.body.style['webkit-BoxSizing'] is undefined.
PASS document.body.style['webkitBox-sizing'] is undefined.
PASS document.body.style['webkitBox-Sizing'] is undefined.
PASS document.body.style['-webkit-box-sizing'] is "border-box"
PASS document.body.style['cssWebkitBorderBoxSizing'] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
